+++ source/Classes/JeaRoleCapabilities.ps1
@@ -1,18 +1,5 @@
-enum Ensure
-{
-    Present
-    Absent
-}
-
-$modulePath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent) -Parent) -ChildPath 'Modules'
-
-Import-Module -Name (Join-Path -Path $modulePath -ChildPath DscResource.Common)
-Import-Module -Name (Join-Path -Path $modulePath -ChildPath (Join-Path -Path JeaDsc.Common -ChildPath JeaDsc.Common.psm1))
-
-$script:localizedData = Get-LocalizedData -DefaultUICulture en-US
-
 [DscResource()]
-class JeaRoleCapabilities
+class JeaRoleCapabilities:RoleCapabilitiesUtility
 {
 
     [DscProperty()]
@@ -97,29 +84,6 @@ class JeaRoleCapabilities
     [DscProperty()]
     [string[]]$AssembliesToLoad
 
-    hidden [boolean] ValidatePath()
-    {
-        $fileObject = [System.IO.FileInfo]::new($this.Path)
-        Write-Verbose -Message "Validating Path: $($fileObject.Fullname)"
-        Write-Verbose -Message "Checking file extension is psrc for: $($fileObject.Fullname)"
-        if ($fileObject.Extension -ne '.psrc')
-        {
-            Write-Verbose -Message "Doesn't have psrc extension for: $($fileObject.Fullname)"
-            return $false
-        }
-
-        Write-Verbose -Message "Checking parent forlder is RoleCapabilities for: $($fileObject.Fullname)"
-        if ($fileObject.Directory.Name -ne 'RoleCapabilities')
-        {
-            Write-Verbose -Message "Parent folder isn't RoleCapabilities for: $($fileObject.Fullname)"
-            return $false
-        }
-
-
-        Write-Verbose -Message 'Path is a valid psrc path. Returning true.'
-        return $true
-    }
-
     [JeaRoleCapabilities] Get()
     {
         $currentState = [JeaRoleCapabilities]::new()
@@ -189,7 +153,7 @@ class JeaRoleCapabilities
                 {
                     if ($functionDefinitionName -notin $desiredState['VisibleFunctions'])
                     {
-                        Write-Verbose -"Function defined but not visible to Role Configuration: $functionDefinitionName"
+                        Write-Verbose "Function defined but not visible to Role Configuration: $functionDefinitionName"
                         Write-Error "Function defined but not visible to Role Configuration: $functionDefinitionName"
                         $invalidConfiguration = $true
                     }